AI config engine: split resolver into a resolving/ package
FlowResolver was one cohesive but large file spanning four distinct concerns.
Split it into a resolving/ package (mirroring loading/): resolver.py stays the
thin wire whose resolve() reads as the pipeline (phases -> dependencies ->
variables -> inline, accumulating errors), delegating to free-function modules:

- phases.py: scheduled-phase resolution + per-phase class/agent/ref checks
- dependencies.py: duplicate/unscheduled/cycle checks (needs no registry)
- variables.py: declarations, defaults, missing; owns _DeclaredVar
- inlining.py: build_resolved_flow + phase/task inlining

Pure structural refactor, no behavior change. Tests move alongside: per-concern
test files under tests/ai/config/resolving/ (shared entry-builders in helpers.py),
with test_resolver.py holding only the wired tests plus two new cases for the
resolve() branches (flow not found, broken flow entry). The enum-invariant guard
moves to test_errors.py.
